+++ src/02_04e/MLP.cpp
@@ -6,7 +6,7 @@ double frand(){
 
 
 // Return a new Perceptron object with the specified number of inputs (+1 for the bias).
-Perceptron::Perceptron(int inputs, double bias){
+Perceptron::Perceptron(size_t inputs, double bias){
 	this->bias = bias;
 	weights.resize(inputs+1);
 	generate(weights.begin(),weights.end(),frand);
