+++ src/02_08/MLP.h
@@ -11,7 +11,7 @@ class Perceptron {
 	public: 
 		std::vector<double> weights;
 		double bias;
-		Perceptron(int inputs, double bias=1.0);
+		Perceptron(size_t inputs, double bias=1.0);
         double run(std::vector<double> x);
 		void set_weights(std::vector<double> w_init);
 		double sigmoid(double x);
@@ -19,13 +19,13 @@ class Perceptron {
 
 class MultiLayerPerceptron {
 	public: 
-		MultiLayerPerceptron(std::vector<int> layers, double bias=1.0, double eta = 0.5);
+		MultiLayerPerceptron(std::vector<size_t> layers, double bias=1.0, double eta = 0.5);
 		void set_weights(std::vector<std::vector<std::vector<double> > > w_init);
 		void print_weights();
 		std::vector<double> run(std::vector<double> x);
 		double bp(std::vector<double> x, std::vector<double> y);
 		
-		std::vector<int> layers;
+		std::vector<size_t> layers;
 		double bias;
 		double eta;
 		std::vector<std::vector<Perceptron> > network;
